cloud_functions/airflow_ae_firewall_allow/requirements.txt
google-api-python-client==1.9.0
google-cloud-container==0.5.0
kubernetes==11.0.0
